+++ src/ChessGame/ChessPiece.java
@@ -16,13 +16,13 @@ import java.util.Map;
  * @param hasMoved has this piece been moved from its original position
  * @author Owan Lazic
  */
-public record ChessPiece(PieceType PieceType, int x, int y, boolean isWhitePiece, boolean hasMoved)
+public record ChessPiece(PieceType PieceType, byte x, byte y, boolean isWhitePiece, boolean hasMoved)
 {
     /**
      * @param board
      * @return a list of arrays containing the x and y position on the board for the piece to move
      */
-    public List<Integer[]> getPossibleMoves(ChessBoard board)
+    public List<byte[]> getPossibleMoves(ChessBoard board)
     {
         return MoveFinder.get(this, board);
     }
@@ -30,27 +30,27 @@ public record ChessPiece(PieceType PieceType, int x, int y, boolean isWhitePiece
 class MoveFinder
 {
     static { precomputeMoves(); }
-    private static Map<PieceType,List<Integer[]>> precomputedMoves;
+    private static Map<PieceType,List<byte[]>> precomputedMoves;
 
 
-    public static List<Integer[]> get(ChessPiece p, ChessBoard board)
+    public static List<byte[]> get(ChessPiece p, ChessBoard board)
     {
-        List<Integer[]> possibleMoves = new ArrayList<>();
+        List<byte[]> possibleMoves = new ArrayList<>();
 
         if (precomputedMoves.containsKey(p.PieceType()))
         {
             possibleMoves = precomputedMoves.get(p.PieceType());
             possibleMoves = possibleMoves.stream() // make relative to piece position and filter out, out of bounds positions
-                    .map(m->new Integer[]{m[0]+p.x(),m[1]+p.y()})
+                    .map(m->new byte[]{(byte) (m[0]+p.x()), (byte) (m[1]+p.y())})
                     .filter(m->m[0] >= 0 && m[1] >= 0)
                     .filter(m->m[0] < 8 && m[1] < 8)
                     .toList();
         }
         else
         {
-            for (int x = 0; x < 8; x++)
-                for (int y = 0; y < 8; y++)
-                    possibleMoves.add(new Integer[]{x, y});
+            for (byte x = 0; x < 8; x++)
+                for (byte y = 0; y < 8; y++)
+                    possibleMoves.add(new byte[]{x, y});
         }
 
         return possibleMoves.stream()
@@ -65,12 +65,12 @@ class MoveFinder
         precomputedMoves.put(
                 PieceType.PAWN,
                 List.of(
-                        new Integer[]{0,1},
-                        new Integer[]{0,-1},
-                        new Integer[]{1,1},
-                        new Integer[]{-1,1},
-                        new Integer[]{0,2},
-                        new Integer[]{0,-2}
+                        new byte[]{0,1},
+                        new byte[]{0,-1},
+                        new byte[]{1,1},
+                        new byte[]{-1,1},
+                        new byte[]{0,2},
+                        new byte[]{0,-2}
                 )
         );
 
@@ -78,16 +78,16 @@ class MoveFinder
                 (
                         PieceType.KING,
                         List.of(
-                                new Integer[]{0,1},
-                                new Integer[]{0,-1},
-                                new Integer[]{1,0},
-                                new Integer[]{-1,0},
-                                new Integer[]{1,1},
-                                new Integer[]{1,-1},
-                                new Integer[]{-1,1},
-                                new Integer[]{-1,-1},
-                                new Integer[]{3,0},
-                                new Integer[]{-2,0}
+                                new byte[]{0,1},
+                                new byte[]{0,-1},
+                                new byte[]{1,0},
+                                new byte[]{-1,0},
+                                new byte[]{1,1},
+                                new byte[]{1,-1},
+                                new byte[]{-1,1},
+                                new byte[]{-1,-1},
+                                new byte[]{3,0},
+                                new byte[]{-2,0}
                         )
                 );
 
@@ -95,55 +95,55 @@ class MoveFinder
                 (
                         PieceType.BISHOP,
                         List.of(
-                                new Integer[]{1,1},
-                                new Integer[]{1,-1},
-                                new Integer[]{-1,1},
-                                new Integer[]{-1,-1},
-                                new Integer[]{2,2},
-                                new Integer[]{2,-2},
-                                new Integer[]{-2,2},
-                                new Integer[]{-2,-2},
-                                new Integer[]{3,3},
-                                new Integer[]{3,-3},
-                                new Integer[]{-3,3},
-                                new Integer[]{-3,-3},
-                                new Integer[]{4,4},
-                                new Integer[]{4,-4},
-                                new Integer[]{-4,4},
-                                new Integer[]{-4,-4},
-                                new Integer[]{5,5},
-                                new Integer[]{5,-5},
-                                new Integer[]{-5,5},
-                                new Integer[]{-5,-5},
-                                new Integer[]{6,6},
-                                new Integer[]{6,-6},
-                                new Integer[]{-6,6},
-                                new Integer[]{-6,-6},
-                                new Integer[]{7,7},
-                                new Integer[]{7,-7},
-                                new Integer[]{-7,7},
-                                new Integer[]{-7,-7},
-                                new Integer[]{8,8},
-                                new Integer[]{8,-8},
-                                new Integer[]{-8,8},
-                                new Integer[]{-8,-8}
+                                new byte[]{1,1},
+                                new byte[]{1,-1},
+                                new byte[]{-1,1},
+                                new byte[]{-1,-1},
+                                new byte[]{2,2},
+                                new byte[]{2,-2},
+                                new byte[]{-2,2},
+                                new byte[]{-2,-2},
+                                new byte[]{3,3},
+                                new byte[]{3,-3},
+                                new byte[]{-3,3},
+                                new byte[]{-3,-3},
+                                new byte[]{4,4},
+                                new byte[]{4,-4},
+                                new byte[]{-4,4},
+                                new byte[]{-4,-4},
+                                new byte[]{5,5},
+                                new byte[]{5,-5},
+                                new byte[]{-5,5},
+                                new byte[]{-5,-5},
+                                new byte[]{6,6},
+                                new byte[]{6,-6},
+                                new byte[]{-6,6},
+                                new byte[]{-6,-6},
+                                new byte[]{7,7},
+                                new byte[]{7,-7},
+                                new byte[]{-7,7},
+                                new byte[]{-7,-7},
+                                new byte[]{8,8},
+                                new byte[]{8,-8},
+                                new byte[]{-8,8},
+                                new byte[]{-8,-8}
                         )
                 );
 
         precomputedMoves.put(
                 PieceType.KNIGHT,
                 List.of(
-                        new Integer[]{1,2},
-                        new Integer[]{-1,2},
+                        new byte[]{1,2},
+                        new byte[]{-1,2},
 
-                        new Integer[]{2,1},
-                        new Integer[]{2,-1},
+                        new byte[]{2,1},
+                        new byte[]{2,-1},
 
-                        new Integer[]{1,-2},
-                        new Integer[]{-1,-2},
+                        new byte[]{1,-2},
+                        new byte[]{-1,-2},
 
-                        new Integer[]{-2,1},
-                        new Integer[]{-2,-1}
+                        new byte[]{-2,1},
+                        new byte[]{-2,-1}
 
                 )
         );
@@ -151,38 +151,38 @@ class MoveFinder
         precomputedMoves.put(
                 PieceType.ROOK,
                 List.of(
-                        new Integer[]{1,0},
-                        new Integer[]{0,1},
-                        new Integer[]{-1,0},
-                        new Integer[]{0,-1},
-                        new Integer[]{2,0},
-                        new Integer[]{0,2},
-                        new Integer[]{-2,0},
-                        new Integer[]{0,-2},
-                        new Integer[]{3,0},
-                        new Integer[]{0,3},
-                        new Integer[]{-3,0},
-                        new Integer[]{0,-3},
-                        new Integer[]{4,0},
-                        new Integer[]{0,4},
-                        new Integer[]{-4,0},
-                        new Integer[]{0,-4},
-                        new Integer[]{5,0},
-                        new Integer[]{0,5},
-                        new Integer[]{-5,0},
-                        new Integer[]{0,-5},
-                        new Integer[]{6,0},
-                        new Integer[]{0,6},
-                        new Integer[]{-6,0},
-                        new Integer[]{0,-6},
-                        new Integer[]{7,0},
-                        new Integer[]{0,7},
-                        new Integer[]{-7,0},
-                        new Integer[]{0,-7},
-                        new Integer[]{8,0},
-                        new Integer[]{0,8},
-                        new Integer[]{-8,0},
-                        new Integer[]{0,-8}
+                        new byte[]{1,0},
+                        new byte[]{0,1},
+                        new byte[]{-1,0},
+                        new byte[]{0,-1},
+                        new byte[]{2,0},
+                        new byte[]{0,2},
+                        new byte[]{-2,0},
+                        new byte[]{0,-2},
+                        new byte[]{3,0},
+                        new byte[]{0,3},
+                        new byte[]{-3,0},
+                        new byte[]{0,-3},
+                        new byte[]{4,0},
+                        new byte[]{0,4},
+                        new byte[]{-4,0},
+                        new byte[]{0,-4},
+                        new byte[]{5,0},
+                        new byte[]{0,5},
+                        new byte[]{-5,0},
+                        new byte[]{0,-5},
+                        new byte[]{6,0},
+                        new byte[]{0,6},
+                        new byte[]{-6,0},
+                        new byte[]{0,-6},
+                        new byte[]{7,0},
+                        new byte[]{0,7},
+                        new byte[]{-7,0},
+                        new byte[]{0,-7},
+                        new byte[]{8,0},
+                        new byte[]{0,8},
+                        new byte[]{-8,0},
+                        new byte[]{0,-8}
                 )
         );
     }
