+++ src/ChessGame/Rules/MoveDiagonalRule.java
@@ -16,7 +16,7 @@ public class MoveDiagonalRule implements MoveRule
     private final Integer numSquares;
     public MoveDiagonalRule(Integer numSquares) {this.numSquares = numSquares;}
     @Override
-    public Boolean isValidMove(ChessBoard board, ChessPiece piece, int toX, int toY)
+    public Boolean isValidMove(ChessBoard board, ChessPiece piece, byte toX, byte toY)
     {
         return Math.abs(piece.x() - toX) == Math.abs(piece.y() - toY)
                 && Math.abs(piece.x() - toX) <= numSquares;
