+++ src/ChessGame/Rules/MoveUpOnlyRule.java
@@ -14,5 +14,5 @@ import ChessGame.ChessPiece;
 public class MoveUpOnlyRule implements MoveRule
 {
     @Override
-    public Boolean isValidMove(ChessBoard board, ChessPiece piece, int toX, int toY) {return piece.isWhitePiece() ? toY > piece.y() : toY < piece.y();}
+    public Boolean isValidMove(ChessBoard board, ChessPiece piece, byte toX, byte toY) {return piece.isWhitePiece() ? toY > piece.y() : toY < piece.y();}
 }
