+++ src/ChessGame/Rules/MoveVerticalRule.java
@@ -15,7 +15,7 @@ public class MoveVerticalRule implements MoveRule
     Integer maxNumberOfSquares;
     public MoveVerticalRule(Integer maxNumberOfSquares) {this.maxNumberOfSquares = maxNumberOfSquares;}
     @Override
-    public Boolean isValidMove(ChessBoard board, ChessPiece piece, int toX, int toY)
+    public Boolean isValidMove(ChessBoard board, ChessPiece piece, byte toX, byte toY)
     {
        return toX == piece.x() && Math.abs(toY - piece.y()) <= maxNumberOfSquares;
     }
